Moved conversion function to utils and switched from BBHPriorDict to PriorDict in utils
